fix: csv import; `sample` without `dashboard` shows case table
