+++ v3/src/components/app.tsx
@@ -1,28 +1,25 @@
 import { observer } from "mobx-react-lite"
-import { getSnapshot } from "mobx-state-tree"
 import React, { useCallback, useEffect } from "react"
+import { kCaseTableTileType } from "./case-table/case-table-defs"
 import { CodapDndContext } from "./codap-dnd-context"
+import { Container } from "./container/container"
+import { createTileOfType } from "./create-tile"
 import { ToolShelf } from "./tool-shelf/tool-shelf"
-import {Container} from "./container"
+import { importV2Document } from "./import-v2-document"
 import { MenuBar } from "./menu-bar/menu-bar"
 import { appState } from "../models/app-state"
 import { addDefaultComponents } from "../models/codap/add-default-content"
-import { createCodapDocument } from "../models/codap/create-codap-document"
 import {gDataBroker} from "../models/data/data-broker"
 import {DataSet, IDataSet, toCanonical} from "../models/data/data-set"
 import { IDocumentModelSnapshot } from "../models/document/document"
-import { wrapSerialization } from "../models/shared/shared-data-utils"
-import { getTileComponentInfo } from "../models/tiles/tile-component-info"
+import { linkTileToDataSet } from "../models/shared/shared-data-utils"
 import { getSharedModelManager } from "../models/tiles/tile-environment"
-import { ITileModel } from "../models/tiles/tile-model"
 import { DocumentContext } from "../hooks/use-document-context"
 import {useDropHandler} from "../hooks/use-drop-handler"
 import { useKeyStates } from "../hooks/use-key-states"
 import { registerTileTypes } from "../register-tile-types"
 import { importSample, sampleData } from "../sample-data"
 import { urlParams } from "../utilities/url-params"
-import { CodapV2Document } from "../v2/codap-v2-document"
-import { importV2Component } from "../v2/codap-v2-tile-importers"
 import t from "../utilities/translation/translate"
 
 import "../models/shared/shared-case-metadata-registration"
@@ -32,9 +29,20 @@ import "./app.scss"
 
 registerTileTypes([])
 
-export function handleImportDataSet(data: IDataSet) {
+interface IImportDataSetOptions {
+  createTableTile?: boolean // default true
+}
+export function handleImportDataSet(data: IDataSet, options?: IImportDataSetOptions) {
   // add data set
-  gDataBroker.addDataSet(data)
+  const { sharedData } = gDataBroker.addDataSet(data)
+  if (sharedData.dataSet && (options?.createTableTile !== false)) {
+    // create the corresponding case table
+    const newTile = createTileOfType(kCaseTableTileType, appState.document.content)
+    if (newTile) {
+      // link the case table to the new data set
+      linkTileToDataSet(newTile.content, sharedData.dataSet)
+    }
+  }
 }
 
 export const App = observer(function App() {
@@ -42,57 +50,14 @@ export const App = observer(function App() {
 
   useKeyStates()
 
-  const _handleImportDataSet = useCallback((data: IDataSet) => {
-    handleImportDataSet(data)
-  }, [])
-
-  const handleImportV2Document = useCallback((v2Document: CodapV2Document) => {
-    const v3Document = createCodapDocument(undefined, { layout: "free" })
-    const sharedModelManager = getSharedModelManager(v3Document)
-    sharedModelManager && gDataBroker.setSharedModelManager(sharedModelManager)
-    // add shared models (data sets and case metadata)
-    v2Document.datasets.forEach((data, key) => {
-      const metadata = v2Document.metadata[key]
-      gDataBroker.addDataAndMetadata(data, metadata)
-    })
-
-    // sort by zIndex so the resulting tiles will be ordered appropriately
-    const v2Components = v2Document.components.slice()
-    v2Components.sort((a, b) => (a.layout.zIndex ?? 0) - (b.layout.zIndex ?? 0))
-
-    // add components
-    const { content } = v3Document
-    const row = content?.firstRow
-    v2Components.forEach(v2Component => {
-      const insertTile = (tile: ITileModel) => {
-        if (row && tile) {
-          const info = getTileComponentInfo(tile.content.type)
-          if (info) {
-            const { left, top, width, height } = v2Component.layout
-            // only apply imported width and height to resizable tiles
-            const _width = !info.isFixedWidth ? { width } : {}
-            const _height = !info?.isFixedHeight ? { height } : {}
-            content?.insertTileInRow(tile, row, { x: left, y: top, ..._width, ..._height })
-          }
-        }
-      }
-      importV2Component({ v2Component, v2Document, sharedModelManager, insertTile })
-    })
-
-    // retrieve document snapshot
-    const docSnapshot = wrapSerialization(v3Document, () => getSnapshot(v3Document))
-    // use document snapshot
-    appState.setDocument(docSnapshot)
-  }, [])
-
   const handleImportV3Document = useCallback((document: IDocumentModelSnapshot) => {
     appState.setDocument(document)
   }, [])
 
   useDropHandler({
     selector: "#app",
-    onImportDataSet: _handleImportDataSet,
-    onImportV2Document: handleImportV2Document,
+    onImportDataSet: handleImportDataSet,
+    onImportV2Document: importV2Document,
     onImportV3Document: handleImportV3Document
   })
 
@@ -120,7 +85,8 @@ export const App = observer(function App() {
         if (sample) {
           try {
             const data = await importSample(sample)
-            handleImportDataSet(data)
+            // show case table if not showing a complete dashboard
+            handleImportDataSet(data, { createTableTile: !isDashboard })
           }
           catch (e) {
             console.warn(`Failed to import sample "${sample}"`)
@@ -144,7 +110,7 @@ export const App = observer(function App() {
       <DocumentContext.Provider value={appState.document.content}>
         <div className="app" data-testid="app">
           <MenuBar/>
-          <ToolShelf content={codapDocument.content}/>
+          <ToolShelf content={codapDocument.content} createTileOfType={createTileOfType}/>
           <Container content={codapDocument.content}/>
         </div>
       </DocumentContext.Provider>
