+++ v3/src/components/container/container.scss
@@ -1,9 +1,9 @@
-@import "../components/vars";
+@import "../../components/vars";
 
 .codap-container {
   width: 100%;
   height: calc(100% - $menu-bar-height - $tool-shelf-height);
-  background: url("../assets/bg-grid-grey-v3.png");
+  background: url("../../assets/bg-grid-grey-v3.png");
   position: relative;
 
   .component-placeholder {
