+++ v3/src/components/container/container.tsx
@@ -1,13 +1,13 @@
 import { clsx } from "clsx"
 import React, { useCallback } from "react"
+import { useContainerDroppable, getDragTileId } from "../../hooks/use-drag-drop"
+import { IDocumentContentModel } from "../../models/document/document-content"
+import { isFreeTileRow } from "../../models/document/free-tile-row"
+import { isMosaicTileRow } from "../../models/document/mosaic-tile-row"
+import { getSharedModelManager } from "../../models/tiles/tile-environment"
+import { urlParams } from "../../utilities/url-params"
 import { FreeTileRowComponent } from "./free-tile-row"
 import { MosaicTileRowComponent } from "./mosaic-tile-row"
-import { IDocumentContentModel } from "../models/document/document-content"
-import { getSharedModelManager } from "../models/tiles/tile-environment"
-import { isFreeTileRow } from "../models/document/free-tile-row"
-import { isMosaicTileRow } from "../models/document/mosaic-tile-row"
-import { getDragTileId, useContainerDroppable } from "../hooks/use-drag-drop"
-import { urlParams } from "../utilities/url-params"
 
 import "./container.scss"
