+++ v3/src/components/container/free-tile-row.scss
@@ -1,4 +1,4 @@
-@import "./vars.scss";
+@import "../vars.scss";
 
 .free-tile-row {
   position: relative;
